Tidying up and refactoring
