+++ src/foundrymq/FoundryMQ.ts
@@ -1,5 +1,5 @@
 import { Client } from './client'
-import { Log } from "./Log"
+import { Log } from "./logs/Log"
 
 export const FoundryMQ = new Client({
     server: '/mq',
