+++ src/foundrymq/asd.ts
@@ -1,8 +1,6 @@
 import { unwrap } from "../helpers/assertions"
-import moduleName from "../helpers/module-name"
 
 import { UpdateActorMessage } from '../../foundrymq/messages/UpdateActorMessage'
-import { KEY, Log } from "./Log"
 import { FoundryMQ } from "./FoundryMQ"
 
 async function processMessage(message: UpdateActorMessage): Promise<void>
@@ -22,89 +20,6 @@ async function processMessage(message: UpdateActorMessage): Promise<void>
         },
     })
 }
-
-declare global
-{
-    namespace ClientSettings
-    {
-        interface Values
-        {
-            'wor.heroLabLog': Array<string>
-        }
-    }
-}
-
-class HeroLabLogViewer extends Application
-{
-    static instance: HeroLabLogViewer | undefined
-
-    static override get defaultOptions(): Application.Options
-    {
-        return {
-            ...super.defaultOptions,
-            template: 'systems/wor/src/foundrymq/asd.hbs',
-            width: 400,
-            height: 400,
-            resizable: true,
-            title: 'Hero Lab Logs',
-        }
-    }
-
-    override _getHeaderButtons()
-    {
-        const buttons = super._getHeaderButtons()
-        if (unwrap(game.user).isGM)
-        {
-            buttons.unshift({
-                label: 'Clear',
-                class: 'clear',
-                icon: 'fas fa-trash',
-                onclick: () => Log.clear(),
-            })
-        }
-        return buttons
-    }
-
-    override getData(): object
-    {
-        return {
-            content: game.settings.get(moduleName, KEY)
-        }
-    }
-}
-
-declare global
-{
-    interface Game
-    {
-        heroLab: {
-            showLogs: () => void
-        }
-    }
-}
-
-Hooks.on('init', function()
-{
-    game.heroLab = {
-        showLogs(): void
-        {
-            HeroLabLogViewer.instance ??= new HeroLabLogViewer()
-            HeroLabLogViewer.instance.render(true)
-        }
-    }
-
-    game.settings.register(moduleName, KEY, {
-        scope: 'world',
-        config: false,
-        default: [],
-        type: Array,
-        onChange: () =>
-        {
-            if (HeroLabLogViewer.instance?.rendered)
-                HeroLabLogViewer.instance.render(true)
-        },
-    })
-})
 
 Hooks.once('ready', function()
 {
