+++ src/foundrymq/logs/Viewer.hbs
